Propagate constness in some arguments.
Reduce manual memory management.

Allow the compiler to take advantage of compile-time knowledge of the
dimension when using the templated form of SPTree.  Especially since,
for us, D=2 or 3, the benefits of loop unrolling can be significant.

Remove blas from the python build.  It isn't actually being used, so
there's no reason to introduce a dependency on it.

Adjust the build settings for the python setup build to be the same as
the Makefile, at least on linux.  Specifically, turn on LTO and
gc-sections.  These make a large difference in binary size, especially
when statically linking libc++.