Add optimizer hints for hot paths.
Also add visibility attributes for internal classes.  This allows for
more aggressive linker garbage collection and optimization.  Turning on
-fvisiblity=hidden globablly would require cython to set visibility
attributes on the entry points it exports, so we're not going to get the
full benefit, unfortunately.

Also change containsPoint comparison to abs(corner-point)>width.
This replaces a second subtract/compare with a single vectorized
bitwise and instruction, and by reducing branch counts also helps CPU
pipelining.