Major build cleanup.
Fix the Makefile dependencies so things get rebuilt when they should be.
Add a `test` target.

Only produce one .so file, which knows how to run both 2d and 3d.

Break out main and the I/O methods into a separate cpp file, which is
not used for the python extension module.

Do not create pointless objects.  The methods are all static.  TSNE is
now a namespace, not a class, and the exported `run` method is wrapped
in `extern "C"` to make linkage easier.

Minor cleanup of includes / using statements.

Because I just couldn't help myself, remove a couple places where
new/delete were being used for no good reason.

Get rid of a few places where NDIMS (the build time constant) and a
run-time variable, which are supposed to have the same value, were both
being used in a method.  This means flowing the template a bit further
up the tree.  There's lots more places this could be done - this only
addresses the cases where not doing so could lead to errors if the
template variable was inconsistent with the run-time variable.

Kill off the last pointless vestiges of the cblas dependency.

Use hidden visibility where appropriate - only export the run() method.
This cuts the size of the .so and allows the compiler to inline more
aggressively, especially since we've also turned on LTO.

Together, all of these changes result in a build with a single 901k
extension module, instead of two 1.9MB extension modules.

Set -std=c++14 in the build.  This is required by some of the cleanups
because of the use of the `nullptr` keyword.  Older gcc will not default
to c++11 or higher, so will fail to build nullptr.
And c++17 doesn't work because
```
include/python2.7/stringobject.h:175:5: error:
      ISO C++17 does not allow 'register' storage class specifier [-Wregister]
    register Py_ssize_t *len    /* pointer to length variable or NULL
    ^~~~~~~~~
```

Update travis build to Xenial container.  This updates gcc to 5.4, which has
the required c++ standards support.